Refactor reactive stream sharing modules
