+++ src/heart/utilities/reactivex/__init__.py
@@ -1,0 +1,5 @@
+"""Helpers for sharing and tuning reactive streams."""
+
+from heart.utilities.reactivex.settings import \
+    StreamShareSettings  # noqa: F401
+from heart.utilities.reactivex.sharing import share_stream  # noqa: F401
